Refactor non-negative CP
