+++ tensorly/decomposition/__init__.py
@@ -3,8 +3,8 @@ The :mod:`tensorly.decomposition` module includes utilities for performing
 tensor decomposition such as CANDECOMP-PARAFAC and Tucker.                                                                                               
 """
 
-from ._cp import (parafac, non_negative_parafac, CP, RandomizedCP,
-                                randomised_parafac, sample_khatri_rao)
+from ._cp import (parafac, CP, RandomizedCP, randomised_parafac, sample_khatri_rao)
+from ._nn_cp import non_negative_parafac
 from ._tucker import tucker, partial_tucker, non_negative_tucker, Tucker
 from .robust_decomposition import robust_pca
 from ._tt import TensorTrain, tensor_train, tensor_train_matrix
